#46 Mockup for quick posts
Pulls 3 items off the top of _data/updates.yml and displays on home
page (new items should be written to the top of the file). The part of
the sentence that isn’t the staffer’s name is linked if a URL is
provided. Would be nice to fix this so name links to staffer’s bio
page, but not sure how to do this when multiple staffers named per
update.