Changes `name` to `title` for people frontmatter.
Ensures that all our content files use `title` to denote the actual
title of a page. Given that the only place the `name` YAML field was
used was on the person layout, it seems better to simply make our
YAML fields uniform instead of trying to account for the absence of a
`title` field in the search results list and in the construction of the
search corpus.

Fixes #527.