Updates nav info arch, w/1 exception...
Changes navigation to match latest plan of pages/org; adjusts /pages YAML, config, CSS to match; adds pages that didn't exist.
NOTE: eventually want first link to be “About” with a submenu to “Charter” (and eventually “Accessibility”), not "Opportunities". I left in the “Opportunities” > students opps setup for now instead, since the demo of dropdown is tied into that and I couldn't figure out in 30 seconds how to edit that