Sub-sub-folder items get shorter URLs...
These use YAML permalinks so they are available at shorter URLs, rather than having the URL display the full file path (e.g. /lami instead of /pages/collaborate/students/student-roles/lami). These also happen to be the same as the links on the current site, meaning no redirects are needed. So we were able to organize the site files into folders nicely so it's easier to find things, but not have that mess up links.