Adds conditional for displaying author name.
We currently use two different fields for the people record name:
  - person.title
  - person.name

Our people records use one or the other. I don't know when this
change or discrepancy happened. We might wish to make these
consistent one day?

This conditional checks for either of these and displays the one
that exists.