Create 2025-04-24-design-jams1 (#1163)
* Create 2025-04-24-design-jams1

adding blog post to personal branch on the repo

* Update and rename 2025-04-24-design-jams1 to 2025-04-24-design-jams1.md

fixed file name, edited the silliness in the title section, refined some of the wordings.

* Update 2025-04-24-design-jams1.md

still playing with the silliness in the title section. needs review by someone who understands the css for the site, and still might need to just be cut

* my update (#1164)

---------

Co-authored-by: Ronda <rag9b@virginia.edu>