Corrects number of recent blog posts list. Fixes #1156.
Updates the posts query on the home page for recent posts
to rely on the `layout` YAML field to reliably exclude "update"
posts. There is currently no reliable way to run a posts query that
explicitly excludes a specific category (e.g. posts not
categorized 'update').

This is a somewhat brittle solution, and relies on us continuing to
use a separate layout for updates in additions to *not* using separate
layouts for other kinds of posts.